[fish] update web commmand to support ./bin/django.sh

+++ .config/fish/functions/web.fish
@@ -1,3 +1,7 @@
 function web --description 'alias web docker compose run --rm web'
-  docker compose run --rm web $argv
+  if test -f ./bin/django.sh
+    ./bin/django.sh $argv
+  else
+    docker compose run --rm web $argv
+  end
 end
